Replace require with import
