+++ index.js
@@ -1,15 +1,21 @@
-require('dotenv').config()
-const { Client, Intents, MessageEmbed } = require('discord.js')
-const ytdl = require('ytdl-core-discord')
-const {
-  joinVoiceChannel,
+import { Client, Intents, MessageEmbed } from 'discord.js'
+import 'dotenv/config'
+import ytdl from "ytdl-core-discord"
+import {
   createAudioPlayer,
   createAudioResource,
   AudioPlayerStatus,
-  StreamType,
-  getVoiceConnection
-} = require('@discordjs/voice')
-const ytsearch = require('youtube-sr')
+  StreamType
+} from '@discordjs/voice'
+import YouTube from 'youtube-sr'
+
+// Controllers
+import parseMessage from './src/controllers/ParseMessage/parse_message.mjs'
+import playerMonitor from './src/controllers/PlayerMonitor/player_monitor.mjs'
+import createConnection from './src/controllers/CreateConnection/create_connection.mjs'
+import connectionMonitor from './src/controllers/ConnectionMonitor/connection_monitor.mjs'
+import clearResults from './src/controllers/ClearResults/clear_results.mjs'
+import chooseTeams from './src/controllers/ChooseTeams/choose-teams.mjs'
 
 const bot = new Client({
   intents: [Intents.FLAGS.GUILDS, Intents.FLAGS.GUILD_MESSAGES, Intents.FLAGS.GUILD_VOICE_STATES]
@@ -20,58 +26,6 @@ bot.login(token)
 var queue = []
 var results = []
 var tempResults = []
-
-function parseMessage(content) {
-  const message = content.toLowerCase().split(' ')
-  const command = message.shift()
-  const search = message.join(' ')
-
-  return { command, search }
-}
-
-function playerMonitor(player) {
-  player.on(AudioPlayerStatus.Idle, () => {
-    console.log('Player status: IDLE\n')
-    if (queue.length === 0) {
-      return
-    }
-
-    queue.shift()
-    if (queue.length > 0) {
-      const { song, message, resource } = queue[0]
-
-      player.play(resource)
-      message.reply(`Lansei a braba: ${song.title} (${song.durationFormatted})`)
-    }
-  })
-
-  player.on(AudioPlayerStatus.Playing, () => console.log(`Player status: PLAYING\nIn queue: ${queue.length - 1}\n`))
-  player.on(AudioPlayerStatus.Buffering, () => console.log('Player status: BUFFERING\n'))
-  player.on(AudioPlayerStatus.Paused, () => console.log('Player status: PAUSED\n'))
-  player.on(AudioPlayerStatus.AutoPaused, () => console.log('Player status: AUTOPAUSED\n'))
-}
-
-function createConnection(channel, guild) {
-  return joinVoiceChannel({
-    channelId: channel.id,
-    guildId: guild.id,
-    adapterCreator: guild.voiceAdapterCreator
-  })
-}
-
-function connectionMonitor(connection, player) {
-  connection.on('disconnected', () => {
-    console.log('\nClearing queue...')
-    queue = []
-    player.stop()
-    player.removeAllListeners()
-    console.log('Disconnected\n')
-  })
-}
-
-function clearResults() {
-  tempResults = []
-}
 
 bot.on('ready', () => {
   bot.user.setActivity('Tarimba')
@@ -96,10 +50,10 @@ bot.on('messageCreate', async message => {
   const player = createAudioPlayer()
   const channel = member.voice.channel
   const connection = createConnection(channel, guild)
-  connectionMonitor(connection, player, channel)
+  connectionMonitor(connection, player, queue)
 
   if (command === ';prox') {
-    clearResults()
+    clearResults(tempResults)
 
     if (queue.length === 0) {
       return console.log('No more musics')
@@ -118,7 +72,7 @@ bot.on('messageCreate', async message => {
         queue.shift()
       }
 
-      playerMonitor(player)
+      playerMonitor(player, queue)
     }
   }
 
@@ -128,7 +82,7 @@ bot.on('messageCreate', async message => {
     }
 
     if (tempResults.length === 0) {
-      tempResults = await ytsearch.YouTube.search(search)
+      tempResults = await YouTube.search(search)
       results = tempResults?.slice(0, 5) ?? []
 
       if (results.length === 0) {
@@ -169,18 +123,8 @@ bot.on('messageCreate', async message => {
         console.log('Added to queue: ', song.title)
         message.reply(`Essa braba foi pra fila: ${song.title} (${song.durationFormatted})`)
       }
-
-      const checkIdle = (player) => {
-        player.on(AudioPlayerStatus.Idle, () => {
-          setTimeout(() => {
-            connection.disconnect()
-          }, 10000)
-        })
-      }
-
-      checkIdle(player)
-      clearResults()
-      playerMonitor(player)
+      clearResults(tempResults)
+      playerMonitor(player, queue)
     } catch (error) {
       queue = []
       connection.disconnect()
@@ -189,34 +133,15 @@ bot.on('messageCreate', async message => {
   }
 
   if (command === ';ajuda') {
-    clearResults()
+    clearResults(tempResults)
 
     const text =
-      'No momento, possuo os seguintes comandos: \n\n`;toca <URL>` ou `;toca <termo-para-busca>`: Toca a música de uma URL ou retorna 5 resultados da busca pelo termo no youtube.\nUse `;toca <número-da-musica>` para tocar a música escolhida. \n\n`;prox`: Toca a próxima música da fila. O comando também serve para parar a música atual se a fila estiver vazia.'
+      'No momento, possuo os seguintes comandos: \n\n`;toca <URL>` ou `;toca <termo-para-busca>`: Toca a música de uma URL ou retorna 5 resultados da busca pelo termo no youtube.\nUse `;toca <número-da-musica>` para tocar a música escolhida. \n\n`;prox`: Toca a próxima música da fila. O comando também serve para parar a música atual se a fila estiver vazia.\n\n`;tira`: Digite um número par de nomes após o comando para sortear estes em dois times diferentes.'
     return message.reply({ content: text, allowedMentions: { repliedUser: true } })
   }
 
   if (command === ';tira') {
-    let resultsArray = search.split(', '), team1 = [], team2 = [], size = resultsArray.length
-
-    if (size % 2 !== 0) {
-      message.reply('O total de jogadores informado deve ser par!')
-    } else {
-      let currentIndex = size, randomIndex;
-      while (currentIndex !== 0) {
-        randomIndex = Math.floor(Math.random() * currentIndex)
-        currentIndex--
-        [resultsArray[currentIndex], resultsArray[randomIndex]] = [resultsArray[randomIndex], resultsArray[currentIndex]]
-      }
-
-      resultsArray.forEach((value, index) => {
-        if (index < (size / 2)) {
-          team1.push(value)
-        } else {
-          team2.push(value)
-        }
-      })
-      message.reply(`Time das putas:${team1} \nTime dos viados:${team2}`)
-    }
+    const teams = chooseTeams(search)
+    message.reply(teams)
   }
 })
