+++ src/controllers/ClearResults/clear_results.mjs
@@ -1,0 +1,3 @@
+export default function clearResults(tempResults) {
+    tempResults = []
+}
