+++ src/controllers/ConnectionMonitor/connection_monitor.mjs
@@ -1,0 +1,9 @@
+export default function connectionMonitor(connection, player, queue) {
+    connection.on('disconnected', () => {
+        console.log('\nClearing queue...')
+        queue = []
+        player.stop()
+        player.removeAllListeners()
+        console.log('Disconnected\n')
+    })
+}
