+++ src/controllers/CreateConnection/create_connection.mjs
@@ -1,0 +1,9 @@
+import { joinVoiceChannel } from "@discordjs/voice"
+
+export default function createConnection(channel, guild) {
+    return joinVoiceChannel({
+        channelId: channel.id,
+        guildId: guild.id,
+        adapterCreator: guild.voiceAdapterCreator
+    })
+}
