+++ src/controllers/ParseMessage/parse_message.mjs
@@ -1,0 +1,7 @@
+export default function parseMessage(content) {
+    const message = content.toLowerCase().split(' ')
+    const command = message.shift()
+    const search = message.join(' ')
+
+    return { command, search }
+}
